refactor(library): move create_part_annotations into combinatorial_library, simplify parse_library
Moves create_part_annotations from imports/library.rs into combinatorial_library.rs
so updates/library.rs no longer reverse-imports from imports/library.rs. Simplifies
parse_library to build SequencePart incrementally instead of via tuple destructuring,
and consumes parts by value instead of cloning. Documents why create_location_accession
stays anchored on a single find_graph_positions result rather than
AccessionSpan::from_resolved_region or a multi-position fan-out: region bounds for
Annotation/Accession-kind regions can extend past the region's own intervaltree, and
at the zero offsets used here, find_graph_positions already funnels through
ResolvedGraph::resolve_anchor's existing single-match collapse before any walk
happens, so there's nothing left to recover by carrying a position slice through.
